Created a simple GUI for processing and viewing flight data in 3D.
It currently accepts data in the Martlet 2 telemetry format and uses the sensor data to test out the state estimation algorithms (currently attitude is not calculated). It then renders this in 3D in realtime. Unfortunately without an actual board I can't properly calibrate the data, so it approximates the biases and scaling factors based on the data collected prior to ignition.

The project is split into two main projects, Avionics and Avionics.GUI, the former being a C library containing code that should be easily ported to run on the STM32 and the latter containing all the C++ code for the GUI.

Avionics.GUI has a few dependencies and so I have included a readme in the directory with the solution file which will guide you through installing them all. It is also using an OpenGL wrapper (Frontier.Engine) which I only started working on 2 weeks ago and so there is a high probability of there being bugs and memory leaks I've not noticed yet.