Created a C implementation of the QUEST quaternion estimator. It is based off the MatLab version I wrote previously - but also has code to handle the singularities caused by rotations of pi about any axis. I've simplified the code as much as can easily be done but I'm sure it can be optimized further if performance becomes a problem.
I am fairly certain it works having checked it with a wide variety of rotated forms of the reference vectors. I haven't checked it outputs the orientation that minimizes Wahba's loss function (which it is designed to do) as this is non-trivial to achieve but I've tried rotating the reference vectors by slightly different amounts and it outputs a plausible rotation somewhere between the two.