Added covariance estimation to the calibration MatLab scripts. This can then be used by the translation kalman filter. Brief tests have shown this gives much better results than the previous value.
Added serial packet timestamp synchronization although for some reason this causes the simulation to lag noticeably behind.

Added a start delimiter of the characters 'SY' to the serial data format.

The bias values for the accelerometer (computed for the QUEST algorithm) are now subtracted from the accelerometer values before being fed into the kalman filter, as opposed to previously where gravity was just subtracted from the y axis.